Rename package to dukii; organize packages; update build/run and ignore data files
