+++ src/main/java/dukii/Dukii.java
@@ -1,15 +1,15 @@
-package duke;
+package dukii;
 
 import java.util.Scanner;
 import java.util.ArrayList;
 import java.io.IOException;
-import duke.ui.Ui;
-import duke.task.TaskList;
-import duke.task.Task;
-import duke.storage.Storage;
-import duke.parser.Parser;
-import duke.command.Command;
-import duke.exception.DukiiException;
+import dukii.ui.Ui;
+import dukii.task.TaskList;
+import dukii.task.Task;
+import dukii.storage.Storage;
+import dukii.parser.Parser;
+import dukii.command.Command;
+import dukii.exception.DukiiException;
 
 public class Dukii {
     private static ArrayList<Task> tasks = new ArrayList<>();
