+++ src/main/java/dukii/command/TodoCommand.java
@@ -1,10 +1,10 @@
-package duke.command;
+package dukii.command;
 
-import duke.task.TaskList;
-import duke.task.ToDo;
-import duke.ui.Ui;
-import duke.storage.Storage;
-import duke.exception.DukiiException;
+import dukii.task.TaskList;
+import dukii.task.ToDo;
+import dukii.ui.Ui;
+import dukii.storage.Storage;
+import dukii.exception.DukiiException;
 
 public class TodoCommand extends Command {
     private final String description;
