+++ src/main/java/dukii/command/UnmarkCommand.java
@@ -1,10 +1,10 @@
-package duke.command;
+package dukii.command;
 
-import duke.task.TaskList;
-import duke.task.Task;
-import duke.ui.Ui;
-import duke.storage.Storage;
-import duke.exception.DukiiException;
+import dukii.task.TaskList;
+import dukii.task.Task;
+import dukii.ui.Ui;
+import dukii.storage.Storage;
+import dukii.exception.DukiiException;
 
 public class UnmarkCommand extends Command {
     private final int index;
