+++ src/main/java/dukii/task/Deadline.java
@@ -1,4 +1,4 @@
-package duke.task;
+package dukii.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
