+++ src/main/java/dukii/task/Event.java
@@ -1,4 +1,4 @@
-package duke.task;
+package dukii.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
