+++ src/main/java/dukii/task/TaskList.java
@@ -1,4 +1,4 @@
-package duke.task;
+package dukii.task;
 
 import java.util.ArrayList;
