+++ src/main/java/dukii/task/ToDo.java
@@ -1,4 +1,4 @@
-package duke.task;
+package dukii.task;
 
 public class ToDo extends Task {
     public ToDo(String description) {
