+++ src/main/java/dukii/ui/Ui.java
@@ -1,4 +1,4 @@
-package duke.ui;
+package dukii.ui;
 
 public class Ui {
     public void showMessage(String message) {
